# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handler.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  internal_handler.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/internal_handler.APIError'
      meta:
        $ref: '#/definitions/internal_handler.Meta'
      success:
        type: boolean
    type: object
  internal_handler.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  internal_handler.notificationResponse:
    properties:
      body:
        type: string
      created_at:
        type: string
      data:
        additionalProperties:
          type: string
        type: object
      id:
        type: string
      is_read:
        type: boolean
      read_at:
        type: string
      title:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
...
  internal_handler.preferencesResponse:
    properties:
      booking_events:
        type: boolean
      email:
        type: boolean
      in_app:
        type: boolean
      promo_events:
        type: boolean
      push:
        type: boolean
      reminders:
        type: boolean
      review_events:
        type: boolean
    type: object
...
  internal_handler.simpleMessageResponse:
    properties:
      message:
        type: string
    type: object
...
  internal_handler.unreadCountResponse:
    properties:
      unread_count:
        type: integer
    type: object
...
  internal_handler.updatePreferencesRequest:
    properties:
      booking_events:
        type: boolean
      email:
        type: boolean
      in_app:
        type: boolean
      promo_events:
        type: boolean
      push:
        type: boolean
      reminders:
        type: boolean
      review_events:
        type: boolean
    type: object
...
info:
  contact: {}
  description: API for the Bani bathhouse marketplace platform.
  title: Bani API
  version: "1.0"
paths:
...
  /my/notification-preferences:
    get:
      description: Returns notification channel and event type preferences for the
        authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.preferencesResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Get notification preferences
      tags:
      - notifications
    put:
      consumes:
      - application/json
      description: Updates notification channel and event type preferences. Only provided
        fields are updated.
      parameters:
      - description: Preferences to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updatePreferencesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.preferencesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Update notification preferences
      tags:
      - notifications
  /my/notifications:
    get:
      description: Returns paginated list of notifications for the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_handler.notificationResponse'
                  type: array
                meta:
                  $ref: '#/definitions/internal_handler.Meta'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: List notifications
      tags:
      - notifications
  /my/notifications/{id}/read:
    patch:
      description: Marks a single notification as read
      parameters:
      - description: Notification ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.simpleMessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - notifications
  /my/notifications/read-all:
    patch:
      description: Marks all notifications as read for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.simpleMessageResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - notifications
  /my/notifications/unread-count:
    get:
      description: Returns the number of unread notifications for the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.unreadCountResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/internal_handler.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Get unread notifications count
      tags:
      - notifications
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"